Added test for division

+++ calculator.cs
@@ -16,7 +16,7 @@ public int multiply (int a, int b)
         return a * b;
     }  
 
-public int division (int a, int b)
+public double division (double a, double b)
     {
         return a / b;
     }
